backend/requirements.txt
fastapi
uvicorn
python-dotenv
whisper
soundcard
numpy
google-generativeai>=0.3.0
fpdf